Fix: Long으로 엔티티 통일 #154

+++ src/main/java/com/hana/ddok/account/dto/AccountSaving100SaveReq.java
@@ -9,7 +9,7 @@ import com.hana.ddok.users.domain.Users;
 import java.time.LocalDate;
 
 public record AccountSaving100SaveReq(
-        Integer payment,
+        Long payment,
         LocalDate endDate,
         Long productsId,
         Long withdrawalAccountId
