+++ src/main/java/com/hana/ddok/budget/dto/BudgetSumFindRes.java
@@ -3,7 +3,7 @@ package com.hana.ddok.budget.dto;
 import com.hana.ddok.budget.domain.Budget;
 
 public record BudgetSumFindRes(
-        Integer sum
+        Long sum
 ) {
     public BudgetSumFindRes(Budget budget) {
         this(
