+++ src/main/java/com/hana/ddok/budget/dto/BudgetSumUpdateReq.java
@@ -1,7 +1,7 @@
 package com.hana.ddok.budget.dto;
 
 public record BudgetSumUpdateReq(
-        Integer sum
+        Long sum
 ) {
 
 }
