+++ src/main/java/com/hana/ddok/depositsaving/dto/DepositsavingFindbyTypeRes.java
@@ -15,10 +15,10 @@ public record DepositsavingFindbyTypeRes(
         LocalDate startDate,
         LocalDate endDate,
         Long initialAmount,
-        Integer payment,
+        Long payment,
         Long targetAmount
         ) {
-    public DepositsavingFindbyTypeRes(Account account, Depositsaving depositsaving, Long initialAmount, Integer payment, Long targetAmount) {
+    public DepositsavingFindbyTypeRes(Account account, Depositsaving depositsaving, Long initialAmount, Long payment, Long targetAmount) {
         this(
                 account.getAccountId(),
                 account.getProducts().getName(),
